Move check_target_region_column_symmetry to separate validation module
